+++ tenable/io/__init__.py
@@ -1,12 +1,14 @@
 '''
 .. autoclass:: TenableIO
 
+.. automodule:: tenable.io.access_groups
 .. automodule:: tenable.io.agent_config
 .. automodule:: tenable.io.agent_exclusions
 .. automodule:: tenable.io.agent_groups
 .. automodule:: tenable.io.agents
 .. automodule:: tenable.io.assets
 .. automodule:: tenable.io.audit_log
+.. automodule:: tenable.io.credentials
 .. automodule:: tenable.io.editor
 .. automodule:: tenable.io.exclusions
 .. automodule:: tenable.io.exports
@@ -14,6 +16,7 @@
 .. automodule:: tenable.io.filters
 .. automodule:: tenable.io.folders
 .. automodule:: tenable.io.groups
+.. automodule:: tenable.io.networks
 .. automodule:: tenable.io.permissions
 .. automodule:: tenable.io.plugins
 .. automodule:: tenable.io.policies
@@ -35,7 +38,7 @@ there may still bee the occasional need to make raw HTTP calls to the IO API.
 The methods listed below aren't run through any naturalization by the library
 aside from the response code checking.  These methods effectively route
 directly into the requests session.  The responses will be Response objects from
-the ``requests`` library.  In all cases, the path is appended to the base 
+the ``requests`` library.  In all cases, the path is appended to the base
 ``url`` paramater that the ``TenableIO`` object was instantiated with.
 
 Example:
@@ -56,12 +59,14 @@ Example:
 import logging, os
 from tenable.errors import UnexpectedValueError
 from tenable.base import APISession
+from .access_groups import AccessGroupsAPI
 from .agent_config import AgentConfigAPI
 from .agent_exclusions import AgentExclusionsAPI
 from .agent_groups import AgentGroupsAPI
 from .agents import AgentsAPI
 from .assets import AssetsAPI
 from .audit_log import AuditLogAPI
+from .credentials import CredentialsAPI
 from .editor import EditorAPI
 from .exclusions import ExclusionsAPI
 from .exports import ExportsAPI
@@ -69,6 +74,7 @@ from .files import FileAPI
 from .filters import FiltersAPI
 from .folders import FoldersAPI
 from .groups import GroupsAPI
+from .networks import NetworksAPI
 from .permissions import PermissionsAPI
 from .plugins import PluginsAPI
 from .policies import PoliciesAPI
@@ -100,7 +106,7 @@ class TenableIO(APISession):
             variable ``TIO_SECRET_KEY`` to acquire the key.
         url (str, optional):
             The base URL that the paths will be appended onto.  The default
-            is ``https://cloud.tenable.com`` 
+            is ``https://cloud.tenable.com``
         retries (int, optional):
             The number of retries to make before failing a request.  The
             default is ``3``.
@@ -108,17 +114,39 @@ class TenableIO(APISession):
             If a 429 response is returned, how much do we want to backoff
             if the response didn't send a Retry-After header.  The default
             backoff is ``1`` second.
-        ua_identity (str, optional):
-            An application identifier to be added into the User-Agent string
-            for the purposes of application identification.
+        vendor (str, optional):
+            The vendor name for the User-Agent string.
+        product (str, optional):
+            The product name for the User-Agent string.
+        build (str, optional):
+            The version or build identifier for the User-Agent string.
 
     Examples:
+        Basic Example:
+
         >>> from tenable.io import TenableIO
         >>> tio = TenableIO('ACCESS_KEY', 'SECRET_KEY')
+
+        Example with proper identification:
+
+        >>> tio = TenableIO('ACCESS_KEY', 'SECRET_KEY',
+        >>>     vendor='Company Name',
+        >>>     product='My Awesome Widget',
+        >>>     build='1.0.0')
+
+        Example with proper identification leveraging environment variables for
+        access and secret keys:
+
+        >>> tio = TenableIO(
+        >>>     vendor='Company Name', product='Widget', build='1.0.0')
     '''
-    
+
     _tzcache = None
     _url = 'https://cloud.tenable.com'
+
+    @property
+    def access_groups(self):
+        return AccessGroupsAPI(self)
 
     @property
     def agent_config(self):
@@ -143,6 +171,10 @@ class TenableIO(APISession):
     @property
     def audit_log(self):
         return AuditLogAPI(self)
+
+    @property
+    def credentials(self):
+        return CredentialsAPI(self)
 
     @property
     def editor(self):
@@ -171,6 +203,10 @@ class TenableIO(APISession):
     @property
     def groups(self):
         return GroupsAPI(self)
+
+    @property
+    def networks(self):
+        return NetworksAPI(self)
 
     @property
     def permissions(self):
@@ -203,7 +239,7 @@ class TenableIO(APISession):
     @property
     def session(self):
         return SessionAPI(self)
-    
+
     @property
     def tags(self):
         return TagsAPI(self)
@@ -224,34 +260,39 @@ class TenableIO(APISession):
     def _tz(self):
         '''
         As we will be using the timezone listing in a lot of parameter checking,
-        we should probably cache the response as a private attribute to speed 
+        we should probably cache the response as a private attribute to speed
         up checking times.
         '''
         if not self._tzcache:
             self._tzcache = self.scans.timezones()
         return self._tzcache
 
-    def __init__(self, access_key=None, secret_key=None, url=None, retries=None, 
-                 backoff=None, ua_identity=None, session=None, proxies=None):
+    def __init__(self, access_key=None, secret_key=None, url=None, retries=None,
+                 backoff=None, ua_identity=None, session=None, proxies=None,
+                 vendor=None, product=None, build=None):
         if access_key:
             self._access_key = access_key
         else:
             self._access_key = os.getenv('TIO_ACCESS_KEY')
-        
+
         if secret_key:
             self._secret_key = secret_key
         else:
             self._secret_key = os.getenv('TIO_SECRET_KEY')
-        
+
         if not self._access_key or not self._secret_key:
             raise UnexpectedValueError('No valid API Keypair Defined')
-        
-        APISession.__init__(self, url, 
-            retries=retries, 
-            backoff=backoff, 
-            ua_identity=ua_identity, 
+
+        super(TenableIO, self).__init__(url,
+            retries=retries,
+            backoff=backoff,
+            ua_identity=ua_identity,
             session=session,
-            proxies=proxies)
+            proxies=proxies,
+            vendor=vendor,
+            product=product,
+            build=build
+        )
 
     def _retry_request(self, response, retries, kwargs):
         '''
@@ -276,7 +317,7 @@ class TenableIO(APISession):
         '''
         Build the session and add the API Keys into the session
         '''
-        APISession._build_session(self, session)
+        super(TenableIO, self)._build_session(session)
         self._session.headers.update({
             'X-APIKeys': 'accessKey={}; secretKey={};'.format(
                 self._access_key, self._secret_key)