+++ tenable/sc/asset_lists.py
@@ -1,6 +1,6 @@
 '''
 asset_lists
-======
+===========
 
 The following methods allow for interaction into the Tenable.sc
 :sc-api:`Assets <Asset.html>` API.  These items are typically seen
@@ -150,7 +150,7 @@ class AssetListAPI(SCEndpoint):
 
         if 'fobj' in kw:
             # Uploads the file object and stores the returned name in filename.
-            kw['filename'] = self._api.files.upload(fobj)
+            kw['filename'] = self._api.files.upload(kw['fobj'])
             del(kw['fobj'])
 
         if 'data_fields' in kw: